requirements.txt
requests
ansicolors
pyyaml
